Use the built in env vars for the api service
This allows us to just pass the same vars that we are already giving
to the other services (PATH_PREFIX and APP_NAME).

Additionally we encode the api version in the source rather than
passing it as a part of a path environment variable.

If we were to pass it in, we run the risk of the code expecting
one version, but the API endpoint providing a different one. This
way, we will always be in sync.