Define the deployment config image in parts and use an image stream
This commit alters how we define the image to deploy for a source
type by using the environment variable "IMAGE_NAMESPACE" to determine
the first part of the image name. Users can also override this value
be specifying a image_namespace key in the collector definition yaml
entry for the source type.

This is necessary because we need an image change trigger to ensure
the collector deployments are re-deployed when a new image becomes
available. And without this, OpenShift was trying to pull the image
from docker.io/buildfactory/topological-inventory-openshift:latest
rather than from the internal registry.